+++ src/struct_and_iterators/Residue.jl
@@ -93,7 +93,7 @@ function zeta(r::Residue)
         iN = findfirst(at -> name(at) == "N", r)
         iC = findfirst(at -> name(at) == "C", r)
         iCB = findfirst(at -> name(at) == "CB", r)
-        zeta = dihedral(coor(r[iCA]), coor(r[iN]), coor(r[iC]), coor(r[iCB]))
+        zeta = dihedral(position(r[iCA]), position(r[iN]), position(r[iC]), position(r[iCB]))
         return zeta
     else
         return NaN
@@ -235,6 +235,35 @@ julia> resname(residues[1])
 
 """
 resname(residue::Residue) = residue.resname
+
+"""
+    residuename(r::Residue)
+
+Returns the full residue name of the residue.
+
+# Positional Arguments
+
+- `residue::Residue`: A `Residue` object.
+
+# Returns
+
+- `String`: The full residue name.
+
+# Example
+
+```jldoctest
+julia> using PDBTools
+
+julia> atoms = read_pdb(PDBTools.SMALLPDB);
+
+julia> residues = collect(eachresidue(atoms));
+
+julia> residuename(residues[1])
+"Alanine"
+```
+
+"""
+residuename(r::Residue) = residuename(resname(r))
 
 """
     residue(residue::Residue)
@@ -545,13 +574,16 @@ end
 #
 # Deal with alternate conformers with the second test
 #
-isprotein(residue::Residue) = 
-    haskey(protein_residues, resname(residue)) || 
+isprotein(residue::Residue) =
+    haskey(protein_residues, resname(residue)) ||
     (length(resname(residue)) == 4 && haskey(protein_residues, @view(resname(residue)[2:4])))
 
-export isprotein
+isnucleoside(residue::Residue) = haskey(nucleoside_residues, resname(residue))
+
+export isprotein, isnucleoside
 export isacidic, isaliphatic, isaromatic, isbasic, ischarged
 export ishydrophobic, isneutral, isnonpolar, ispolar
+export ispurine, ispyrimidine
 export iswater
 
 isacidic(r::Residue) = isprotein(r) && protein_residues[r.resname].type == "Acidic"
@@ -573,6 +605,11 @@ isneutral(atom::Atom) = isprotein(atom) && protein_residues[atom.resname].charge
 ishydrophobic(atom::Atom) = isprotein(atom) && protein_residues[atom.resname].hydrophobic
 ispolar(atom::Atom) = isprotein(atom) && protein_residues[atom.resname].polar
 isnonpolar(atom::Atom) = isprotein(atom) && !ispolar(atom)
+
+ispurine(r::Residue) = isnucleoside(r) && nucleoside_residues[r.resname].type == "Purine"
+ispyrimidine(r::Residue) = isnucleoside(r) && nucleoside_residues[r.resname].type == "Pyrimidine"
+ispurine(atom::Atom) = isnucleoside(atom) && nucleoside_residues[atom.resname].type == "Purine"
+ispyrimidine(atom::Atom) = isnucleoside(atom) && nucleoside_residues[atom.resname].type == "Pyrimidine"
 
 const water_residues = ["HOH", "OH2", "TIP3", "TIP3P", "TIP4P", "TIP5P", "TIP7P", "SPC", "SPCE"]
 iswater(r::Residue; water_residues=water_residues) = r.resname in water_residues